+++ packages/terra-application-utility/src/utility/_UtilityButton.jsx
@@ -7,7 +7,7 @@ import IconChevronDown from 'terra-icon/lib/icon/IconChevronDown';
 import IconChevronRight from 'terra-icon/lib/icon/IconChevronRight';
 import UtilityMenu from './_UtilityMenu';
 import Utils from '../Utils';
-import styles from './_UtilityButton.scss';
+import styles from './_UtilityButton.module.scss';
 
 const cx = classNames.bind(styles);