+++ packages/terra-brand-footer/src/terra-dev-site/doc/brand-footer/BrandFooter.doc.jsx
@@ -10,6 +10,8 @@ import BrandFooterSrc from '!raw-loader!../../../../src/BrandFooter';
 // Example Files
 import DefaultBrandFooter from '../example/DefaultBrandFooter';
 import DefaultBrandFooterSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/DefaultBrandFooter.jsx';
+import VerticalBrandFooter from '../example/VerticalBrandFooter';
+import VerticalBrandFooterSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/VerticalBrandFooter.jsx';
 /* eslint-enabled import/no-webpack-loader-syntax, import/first, import/extensions, import/no-unresolved, import/no-duplicates */
 
 const DocPage = () => (
@@ -22,6 +24,11 @@ const DocPage = () => (
         title: 'BrandFooter - Default',
         example: <DefaultBrandFooter />,
         source: DefaultBrandFooterSrc,
+      },
+      {
+        title: 'BrandFooter - Vertical Navigation Bar',
+        example: <VerticalBrandFooter />,
+        source: VerticalBrandFooterSrc,
       },
     ]}
     propsTables={[