+++ packages/terra-brand-footer/src/terra-dev-site/doc/example/DefaultBrandFooter.jsx
@@ -3,13 +3,24 @@ import React from 'react';
 // eslint-disable-next-line import/no-extraneous-dependencies, import/no-unresolved, import/extensions
 import BrandFooter from 'terra-brand-footer/lib/BrandFooter';
 
-export default () =>
-  (<BrandFooter
-    links={[
-      { text: 'Terra UI', href: 'http://terra-ui.com/' },
-      { text: 'Terra UI Components', href: 'http://terra-ui.com/static/#/site/components' },
-      { text: 'Cerner Home', href: 'https://www.cerner.com/' },
-      { text: 'Cerner Code', href: 'https://code.cerner.com/', target: '_blank' },
+export default () => (
+  <BrandFooter
+    isVertical={false}
+    sections={[
+      {
+        headerText: 'Terra UI Links',
+        links: [
+          { text: 'Terra UI', href: 'http://terra-ui.com/' },
+          { text: 'Terra UI Components', href: 'http://terra-ui.com/static/#/site/components' },
+        ],
+      },
+      {
+        headerText: 'Cerner Links',
+        links: [
+          { text: 'Cerner Home', href: 'https://www.cerner.com/' },
+          { text: 'Cerner Code', href: 'https://code.cerner.com/', target: '_blank' },
+        ],
+      },
     ]}
     contentLeft={
       <a