+++ packages/terra-dialog-modal/src/DialogModal.jsx
@@ -3,7 +3,7 @@ import AbstractModal from 'terra-abstract-modal';
 import PropTypes from 'prop-types';
 import classNames from 'classnames/bind';
 import 'terra-base/lib/baseStyles';
-import styles from './DialogModal.scss';
+import styles from './DialogModal.module.scss';
 
 const cx = classNames.bind(styles);